feat: upgrade to webpack module federation

+++ gatsby-node.js
@@ -5,8 +5,8 @@
  */
 const path = require("path")
 const fs = require("fs-extra")
-const hiveConfig = require("./src/gatsby/content/hive-config.json")
-const indexJSON = require("./src/gatsby/content/index.json")
+const hiveConfig = require("./content/hive-config.json")
+const indexJSON = require("./content/index.json")
 
 // preserve other hive's build files while also deleting the hive build files
 // corresponding to the currently running build
@@ -14,11 +14,16 @@ exports.onPreInit = () => {
   if (process.argv[2] === "build") {
     const buildPath = path.join(__dirname, "public", hiveConfig.id)
     if (fs.existsSync(buildPath)) fs.removeSync(buildPath, { recursive: true })
+    if (fs.existsSync(path.join(__dirname, "public-tmp")))
+      fs.removeSync(path.join(__dirname, "public-tmp"), { recursive: true })
 
-    fs.renameSync(
-      path.join(__dirname, "public"),
-      path.join(__dirname, "public-tmp")
-    )
+    fs.mkdirSync(path.join(__dirname, "public-tmp"))
+
+    if (fs.existsSync(path.join(__dirname, "public")))
+      fs.renameSync(
+        path.join(__dirname, "public"),
+        path.join(__dirname, "public-tmp")
+      )
   }
 }
 
@@ -36,13 +41,13 @@ exports.onPostBuild = function () {
   )
 }
 
-exports.onCreateBabelConfig = ({ actions }) => {
+exports.onCreateBabelConfig = p => {
   if (process.env.NODE_ENV !== "development") {
-    actions.setBabelPlugin({
+    p.actions.setBabelPlugin({
       name: "@babel/plugin-transform-regenerator",
       options: {},
     })
-    actions.setBabelPlugin({
+    p.actions.setBabelPlugin({
       name: "@babel/plugin-transform-runtime",
       options: {},
     })
@@ -54,15 +59,15 @@ exports.createPages = async ({ actions }) => {
 
   createPage({
     path: "/",
-    component: require.resolve("./src/gatsby/src/templates/home"),
+    component: require.resolve("./src/templates/home"),
   })
 
   indexJSON.forEach(element => {
     createPage({
       path: `/story/${element.slug}`,
-      component: require.resolve("./src/gatsby/src/templates/story"),
+      component: require.resolve("./src/templates/story"),
       context: {
-        pageContent: require(`./src/gatsby/content/${element.slug}`),
+        pageContent: require(`./content/${element.slug}`),
       },
     })
   })
